requirements.txt.txt
streamlit
scikit-learn
pandas
numpy
joblib
orange3
orange3-text
nltk